Split gated method support
